rename Header to HeaderTitle
